+++ genesyscloud/architect_user_prompt/genesyscloud_architect_user_prompt_init_test.go
@@ -1,0 +1,57 @@
+package architect_user_prompt
+
+import (
+	"sync"
+	"terraform-provider-genesyscloud/genesyscloud"
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+)
+
+// providerDataSources holds a map of all registered datasources
+var providerDataSources map[string]*schema.Resource
+
+// providerResources holds a map of all registered resources
+var providerResources map[string]*schema.Resource
+
+type registerTestInstance struct {
+	resourceMapMutex   sync.RWMutex
+	datasourceMapMutex sync.RWMutex
+}
+
+// registerTestResources registers all resources used in the tests
+func (r *registerTestInstance) registerTestResources() {
+	r.resourceMapMutex.Lock()
+	defer r.resourceMapMutex.Unlock()
+
+	providerResources[resource_name] = ResourceArchitectUserPrompt()
+	providerResources["genesyscloud_auth_division"] = genesyscloud.ResourceAuthDivision()
+}
+
+// registerTestDataSources registers all data sources used in the tests.
+func (r *registerTestInstance) registerTestDataSources() {
+	r.datasourceMapMutex.Lock()
+	defer r.datasourceMapMutex.Unlock()
+
+	providerDataSources[resource_name] = DataSourceUserPrompt()
+}
+
+// initTestResources initializes all test resources and data sources.
+func initTestResources() {
+	providerDataSources = make(map[string]*schema.Resource)
+	providerResources = make(map[string]*schema.Resource)
+
+	regInstance := &registerTestInstance{}
+
+	regInstance.registerTestResources()
+	regInstance.registerTestDataSources()
+}
+
+// TestMain is a "setup" function called by the testing framework when run the test
+func TestMain(m *testing.M) {
+	// Run setup function before starting the test suite for the architect_user_prompt package
+	initTestResources()
+
+	// Run the test suite for the architect_user_prompt package
+	m.Run()
+}
