+++ genesyscloud/architect_user_prompt/resource_genesyscloud_architect_user_prompt.go
@@ -1,4 +1,4 @@
-package genesyscloud
+package architect_user_prompt
 
 import (
 	"bytes"
@@ -26,701 +26,8 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/mypurecloud/platform-client-sdk-go/v125/platformclientv2"
 )
-
-type PromptAudioData struct {
-	Language string
-	FileName string
-	MediaUri string
-}
-
-type UserPromptStruct struct {
-	ResourceID  string
-	Name        string
-	Description string
-	Resources   []*UserPromptResourceStruct
-}
-
-type UserPromptResourceStruct struct {
-	Language        string
-	Tts_string      string
-	Text            string
-	Filename        string
-	FileContentHash string
-}
-
-var userPromptResource = &schema.Resource{
-	Schema: map[string]*schema.Schema{
-		"language": {
-			Description: "Language for the prompt resource. (eg. en-us)",
-			Type:        schema.TypeString,
-			Required:    true,
-			ValidateFunc: validation.StringInSlice([]string{
-				"af",
-				"af-na",
-				"af-za",
-				"agq",
-				"agq-cm",
-				"ak",
-				"ak-gh",
-				"am",
-				"am-et",
-				"ar",
-				"ar-001",
-				"ar-ae",
-				"ar-bh",
-				"ar-dj",
-				"ar-dz",
-				"ar-eg",
-				"ar-eh",
-				"ar-er",
-				"ar-il",
-				"ar-iq",
-				"ar-jo",
-				"ar-km",
-				"ar-kw",
-				"ar-lb",
-				"ar-ly",
-				"ar-ma",
-				"ar-mr",
-				"ar-om",
-				"ar-ps",
-				"ar-qa",
-				"ar-sa",
-				"ar-sd",
-				"ar-so",
-				"ar-sy",
-				"ar-td",
-				"ar-tn",
-				"ar-ye",
-				"arn-cl",
-				"as",
-				"as-in",
-				"asa",
-				"asa-tz",
-				"az",
-				"az-cyrl",
-				"az-cyrl-az",
-				"az-latn",
-				"az-latn-az",
-				"ba-ru",
-				"bas",
-				"bas-cm",
-				"be",
-				"be-by",
-				"bem",
-				"bem-zm",
-				"bez",
-				"bez-tz",
-				"bg",
-				"bg-bg",
-				"bm",
-				"bm-ml",
-				"bn",
-				"bn-bd",
-				"bn-in",
-				"bo",
-				"bo-cn",
-				"bo-in",
-				"br",
-				"br-fr",
-				"bs",
-				"bs-cyrl",
-				"bs-cyrl-ba",
-				"bs-latn",
-				"bs-latn-ba",
-				"ca",
-				"ca-ad",
-				"ca-es",
-				"cgg",
-				"cgg-ug",
-				"chr",
-				"chr-us",
-				"co-fr",
-				"cs",
-				"cs-cz",
-				"cy",
-				"cy-gb",
-				"da",
-				"da-dk",
-				"dav",
-				"dav-ke",
-				"de",
-				"de-at",
-				"de-be",
-				"de-ch",
-				"de-de",
-				"de-li",
-				"de-lu",
-				"dje",
-				"dje-ne",
-				"dsb-de",
-				"dua",
-				"dua-cm",
-				"dv-mv",
-				"dyo",
-				"dyo-sn",
-				"dz",
-				"dz-bt",
-				"ebu",
-				"ebu-ke",
-				"ee",
-				"ee-gh",
-				"ee-tg",
-				"el",
-				"el-cy",
-				"el-gr",
-				"en",
-				"en-029",
-				"en-150",
-				"en-ag",
-				"en-as",
-				"en-au",
-				"en-bb",
-				"en-be",
-				"en-bm",
-				"en-bs",
-				"en-bw",
-				"en-bz",
-				"en-ca",
-				"en-cm",
-				"en-dm",
-				"en-fj",
-				"en-fm",
-				"en-gb",
-				"en-gd",
-				"en-gg",
-				"en-gh",
-				"en-gi",
-				"en-gm",
-				"en-gu",
-				"en-gy",
-				"en-hk",
-				"en-ie",
-				"en-im",
-				"en-in",
-				"en-je",
-				"en-jm",
-				"en-ke",
-				"en-ki",
-				"en-kn",
-				"en-ky",
-				"en-lc",
-				"en-lr",
-				"en-ls",
-				"en-mg",
-				"en-mh",
-				"en-mp",
-				"en-mt",
-				"en-mu",
-				"en-mw",
-				"en-my",
-				"en-na",
-				"en-ng",
-				"en-nz",
-				"en-pg",
-				"en-ph",
-				"en-pk",
-				"en-pr",
-				"en-pw",
-				"en-sb",
-				"en-sc",
-				"en-sg",
-				"en-sl",
-				"en-ss",
-				"en-sz",
-				"en-tc",
-				"en-to",
-				"en-tt",
-				"en-tz",
-				"en-ug",
-				"en-um",
-				"en-us",
-				"en-vc",
-				"en-vg",
-				"en-vi",
-				"en-vu",
-				"en-ws",
-				"en-za",
-				"en-zm",
-				"en-zw",
-				"eo",
-				"es",
-				"es-419",
-				"es-ar",
-				"es-bo",
-				"es-cl",
-				"es-co",
-				"es-cr",
-				"es-cu",
-				"es-do",
-				"es-ea",
-				"es-ec",
-				"es-es",
-				"es-gq",
-				"es-gt",
-				"es-hn",
-				"es-ic",
-				"es-mx",
-				"es-ni",
-				"es-pa",
-				"es-pe",
-				"es-ph",
-				"es-pr",
-				"es-py",
-				"es-sv",
-				"es-us",
-				"es-uy",
-				"es-ve",
-				"et",
-				"et-ee",
-				"eu",
-				"eu-es",
-				"ewo",
-				"ewo-cm",
-				"fa",
-				"fa-af",
-				"fa-ir",
-				"ff",
-				"ff-sn",
-				"fi",
-				"fi-fi",
-				"fil",
-				"fil-ph",
-				"fo",
-				"fo-fo",
-				"fr",
-				"fr-be",
-				"fr-bf",
-				"fr-bi",
-				"fr-bj",
-				"fr-bl",
-				"fr-ca",
-				"fr-cd",
-				"fr-cf",
-				"fr-cg",
-				"fr-ch",
-				"fr-ci",
-				"fr-cm",
-				"fr-dj",
-				"fr-dz",
-				"fr-fr",
-				"fr-ga",
-				"fr-gf",
-				"fr-gn",
-				"fr-gp",
-				"fr-gq",
-				"fr-ht",
-				"fr-km",
-				"fr-lu",
-				"fr-ma",
-				"fr-mc",
-				"fr-mf",
-				"fr-mg",
-				"fr-ml",
-				"fr-mq",
-				"fr-mr",
-				"fr-mu",
-				"fr-nc",
-				"fr-ne",
-				"fr-pf",
-				"fr-re",
-				"fr-rw",
-				"fr-sc",
-				"fr-sn",
-				"fr-sy",
-				"fr-td",
-				"fr-tg",
-				"fr-tn",
-				"fr-vu",
-				"fr-yt",
-				"fy-nl",
-				"ga",
-				"ga-ie",
-				"gd-gb",
-				"gl",
-				"gl-es",
-				"gsw",
-				"gsw-ch",
-				"gsw-fr",
-				"gu",
-				"gu-in",
-				"guz",
-				"guz-ke",
-				"gv",
-				"gv-gb",
-				"ha",
-				"ha-latn",
-				"ha-latn-gh",
-				"ha-latn-ne",
-				"ha-latn-ng",
-				"ha-ng",
-				"haw",
-				"haw-us",
-				"he",
-				"he-il",
-				"hi",
-				"hi-in",
-				"hr",
-				"hr-ba",
-				"hr-hr",
-				"hsb-de",
-				"hu",
-				"hu-hu",
-				"hy",
-				"hy-am",
-				"id",
-				"id-id",
-				"ig",
-				"ig-ng",
-				"ii",
-				"ii-cn",
-				"is",
-				"is-is",
-				"it",
-				"it-ch",
-				"it-it",
-				"it-sm",
-				"ja",
-				"ja-jp",
-				"jgo",
-				"jgo-cm",
-				"jmc",
-				"jmc-tz",
-				"ka",
-				"ka-ge",
-				"kab",
-				"kab-dz",
-				"kam",
-				"kam-ke",
-				"kde",
-				"kde-tz",
-				"kea",
-				"kea-cv",
-				"khq",
-				"khq-ml",
-				"ki",
-				"ki-ke",
-				"kk",
-				"kk-cyrl",
-				"kk-cyrl-kz",
-				"kk-kz",
-				"kl",
-				"kl-gl",
-				"kln",
-				"kln-ke",
-				"km",
-				"km-kh",
-				"kn",
-				"kn-in",
-				"ko",
-				"ko-kp",
-				"ko-kr",
-				"kok",
-				"kok-in",
-				"ks",
-				"ks-arab",
-				"ks-arab-in",
-				"ksb",
-				"ksb-tz",
-				"ksf",
-				"ksf-cm",
-				"kw",
-				"kw-gb",
-				"ky-kg",
-				"lag",
-				"lag-tz",
-				"lb-lu",
-				"lg",
-				"lg-ug",
-				"ln",
-				"ln-ao",
-				"ln-cd",
-				"ln-cf",
-				"ln-cg",
-				"lo",
-				"lo-la",
-				"lt",
-				"lt-lt",
-				"lu",
-				"lu-cd",
-				"luo",
-				"luo-ke",
-				"luy",
-				"luy-ke",
-				"lv",
-				"lv-lv",
-				"mas",
-				"mas-ke",
-				"mas-tz",
-				"mer",
-				"mer-ke",
-				"mfe",
-				"mfe-mu",
-				"mg",
-				"mg-mg",
-				"mgh",
-				"mgh-mz",
-				"mgo",
-				"mgo-cm",
-				"mi-nz",
-				"mk",
-				"mk-mk",
-				"ml",
-				"ml-in",
-				"mn-cn",
-				"mn-mn",
-				"moh-ca",
-				"mr",
-				"mr-in",
-				"ms",
-				"ms-bn",
-				"ms-my",
-				"ms-sg",
-				"mt",
-				"mt-mt",
-				"mua",
-				"mua-cm",
-				"my",
-				"my-mm",
-				"naq",
-				"naq-na",
-				"nb",
-				"nb-no",
-				"nd",
-				"nd-zw",
-				"ne",
-				"ne-in",
-				"ne-np",
-				"nl",
-				"nl-aw",
-				"nl-be",
-				"nl-cw",
-				"nl-nl",
-				"nl-sr",
-				"nmg",
-				"nmg-cm",
-				"nn",
-				"nn-no",
-				"nso-za",
-				"nus",
-				"nus-sd",
-				"nyn",
-				"nyn-ug",
-				"oc-fr",
-				"om",
-				"om-et",
-				"om-ke",
-				"or",
-				"or-in",
-				"pa",
-				"pa-arab",
-				"pa-arab-pk",
-				"pa-guru",
-				"pa-guru-in",
-				"pa-in",
-				"pl",
-				"pl-pl",
-				"prs-af",
-				"ps",
-				"ps-af",
-				"pt",
-				"pt-ao",
-				"pt-br",
-				"pt-cv",
-				"pt-gw",
-				"pt-mo",
-				"pt-mz",
-				"pt-pt",
-				"pt-st",
-				"pt-tl",
-				"qut-gt",
-				"quz-bo",
-				"quz-ec",
-				"quz-pe",
-				"rm",
-				"rm-ch",
-				"rn",
-				"rn-bi",
-				"ro",
-				"ro-md",
-				"ro-ro",
-				"rof",
-				"rof-tz",
-				"ru",
-				"ru-by",
-				"ru-kg",
-				"ru-kz",
-				"ru-md",
-				"ru-ru",
-				"ru-ua",
-				"rw",
-				"rw-rw",
-				"rwk",
-				"rwk-tz",
-				"sa-in",
-				"sah-ru",
-				"saq",
-				"saq-ke",
-				"sbp",
-				"sbp-tz",
-				"se-fi",
-				"se-no",
-				"se-se",
-				"seh",
-				"seh-mz",
-				"ses",
-				"ses-ml",
-				"sg",
-				"sg-cf",
-				"shi",
-				"shi-latn",
-				"shi-latn-ma",
-				"shi-tfng",
-				"shi-tfng-ma",
-				"si",
-				"si-lk",
-				"sk",
-				"sk-sk",
-				"sl",
-				"sl-si",
-				"sma-no",
-				"sma-se",
-				"smj-no",
-				"smj-se",
-				"smn-fi",
-				"sms-fi",
-				"sn",
-				"sn-zw",
-				"so",
-				"so-dj",
-				"so-et",
-				"so-ke",
-				"so-so",
-				"sq",
-				"sq-al",
-				"sq-mk",
-				"sr",
-				"sr-cyrl",
-				"sr-cyrl-ba",
-				"sr-cyrl-cs",
-				"sr-cyrl-me",
-				"sr-cyrl-rs",
-				"sr-latn",
-				"sr-latn-ba",
-				"sr-latn-cs",
-				"sr-latn-me",
-				"sr-latn-rs",
-				"sv",
-				"sv-fi",
-				"sv-se",
-				"sw",
-				"sw-ke",
-				"sw-tz",
-				"sw-ug",
-				"swc",
-				"swc-cd",
-				"syr-sy",
-				"ta",
-				"ta-in",
-				"ta-lk",
-				"ta-my",
-				"ta-sg",
-				"te",
-				"te-in",
-				"teo",
-				"teo-ke",
-				"teo-ug",
-				"tg-tj",
-				"th",
-				"th-th",
-				"ti",
-				"ti-er",
-				"ti-et",
-				"tk-tm",
-				"tn-za",
-				"to",
-				"to-to",
-				"tr",
-				"tr-cy",
-				"tr-tr",
-				"tt-ru",
-				"twq",
-				"twq-ne",
-				"tzm",
-				"tzm-dz",
-				"tzm-latn",
-				"tzm-latn-ma",
-				"ug-cn",
-				"uk",
-				"uk-ua",
-				"ur",
-				"ur-in",
-				"ur-pk",
-				"uz",
-				"uz-arab",
-				"uz-arab-af",
-				"uz-cyrl",
-				"uz-cyrl-uz",
-				"uz-latn",
-				"uz-latn-uz",
-				"vai",
-				"vai-latn",
-				"vai-latn-lr",
-				"vai-vaii",
-				"vai-vaii-lr",
-				"vi",
-				"vi-vn",
-				"vun",
-				"vun-tz",
-				"wo-sn",
-				"yav",
-				"yav-cm",
-				"yo",
-				"yo-ng",
-				"zh",
-				"zh-cn",
-				"zh-hans",
-				"zh-hans-cn",
-				"zh-hans-hk",
-				"zh-hans-mo",
-				"zh-hans-sg",
-				"zh-hant",
-				"zh-hant-hk",
-				"zh-hant-mo",
-				"zh-hant-tw",
-				"zh-hk",
-				"zh-mo",
-				"zh-sg",
-				"zh-tw",
-				"zu",
-				"zu-za",
-			}, false),
-		},
-		"tts_string": {
-			Description: "Text to Speech (TTS) value for the prompt.",
-			Type:        schema.TypeString,
-			Optional:    true,
-		},
-		"text": {
-			Description: "Text value for the prompt.",
-			Type:        schema.TypeString,
-			Optional:    true,
-		},
-		"filename": {
-			Description: "Path or URL to the file to be uploaded as prompt.",
-			Type:        schema.TypeString,
-			Optional:    true,
-		},
-		"file_content_hash": {
-			Description: "Hash value of the audio file content. Used to detect changes. Only required when uploading a local audio file.",
-			Type:        schema.TypeString,
-			Optional:    true,
-		},
-	},
-}
 
 func getAllUserPrompts(_ context.Context, clientConfig *platformclientv2.Configuration) (resourceExporter.ResourceIDMetaMap, diag.Diagnostics) {
 	resources := make(resourceExporter.ResourceIDMetaMap)
@@ -743,53 +50,6 @@ func getAllUserPrompts(_ context.Context, clientConfig *platformclientv2.Configu
 	}
 
 	return resources, nil
-}
-
-func ArchitectUserPromptExporter() *resourceExporter.ResourceExporter {
-	return &resourceExporter.ResourceExporter{
-		GetResourcesFunc: provider.GetAllWithPooledClient(getAllUserPrompts),
-		RefAttrs:         map[string]*resourceExporter.RefAttrSettings{}, // No references
-		CustomFileWriter: resourceExporter.CustomFileWriterSettings{
-			RetrieveAndWriteFilesFunc: ArchitectPromptAudioResolver,
-			SubDirectory:              "audio_prompts",
-		},
-	}
-}
-
-func ResourceArchitectUserPrompt() *schema.Resource {
-	return &schema.Resource{
-		Description: "Genesys Cloud User Audio Prompt",
-
-		CreateContext: provider.CreateWithPooledClient(createUserPrompt),
-		ReadContext:   provider.ReadWithPooledClient(readUserPrompt),
-		UpdateContext: provider.UpdateWithPooledClient(updateUserPrompt),
-		DeleteContext: provider.DeleteWithPooledClient(deleteUserPrompt),
-		Importer: &schema.ResourceImporter{
-			StateContext: schema.ImportStatePassthroughContext,
-		},
-		SchemaVersion: 1,
-		Schema: map[string]*schema.Schema{
-			"name": {
-				Description: "Name of the user audio prompt. Note: If the name of the user prompt is changed, this will cause the Prompt to be dropped and recreated with a new ID. This will generate a new ID for the prompt and will invalidate any Architect flows referencing it. ",
-				Type:        schema.TypeString,
-				Required:    true,
-				ForceNew:    true,
-			},
-			"description": {
-				Description: "Description of the user audio prompt.",
-				Type:        schema.TypeString,
-				Optional:    true,
-			},
-			"resources": {
-				Description: "Audio of TTS resources for the audio prompt.",
-				Type:        schema.TypeSet,
-				Optional:    true,
-				Computed:    true,
-				ConfigMode:  schema.SchemaConfigModeAttr,
-				Elem:        userPromptResource,
-			},
-		},
-	}
 }
 
 func createUserPrompt(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
