+++ genesyscloud/architect_user_prompt/resource_genesyscloud_architect_user_prompt_test.go
@@ -1,4 +1,4 @@
-package genesyscloud
+package architect_user_prompt
 
 import (
 	"fmt"
