+++ genesyscloud/resource_genesyscloud_init.go
@@ -18,7 +18,6 @@ func registerDataSources(l registrar.Registrar) {
 	l.RegisterDataSource("genesyscloud_location", DataSourceLocation())
 	l.RegisterDataSource("genesyscloud_auth_division_home", DataSourceAuthDivisionHome())
 	l.RegisterDataSource("genesyscloud_architect_schedules", DataSourceSchedule())
-	l.RegisterDataSource("genesyscloud_architect_user_prompt", dataSourceUserPrompt())
 	l.RegisterDataSource("genesyscloud_auth_division", dataSourceAuthDivision())
 	l.RegisterDataSource("genesyscloud_auth_division_home", DataSourceAuthDivisionHome())
 	l.RegisterDataSource("genesyscloud_group", DataSourceGroup())
@@ -50,7 +49,6 @@ func registerResources(l registrar.Registrar) {
 	l.RegisterResource("genesyscloud_routing_queue", ResourceRoutingQueue())
 	l.RegisterResource("genesyscloud_location", ResourceLocation())
 	l.RegisterResource("genesyscloud_architect_schedules", ResourceArchitectSchedules())
-	l.RegisterResource("genesyscloud_architect_user_prompt", ResourceArchitectUserPrompt())
 	l.RegisterResource("genesyscloud_auth_division", ResourceAuthDivision())
 	l.RegisterResource("genesyscloud_group", ResourceGroup())
 	l.RegisterResource("genesyscloud_idp_adfs", ResourceIdpAdfs())
@@ -91,7 +89,6 @@ func registerResources(l registrar.Registrar) {
 
 func registerExporters(l registrar.Registrar) {
 	l.RegisterExporter("genesyscloud_architect_schedules", ArchitectSchedulesExporter())
-	l.RegisterExporter("genesyscloud_architect_user_prompt", ArchitectUserPromptExporter())
 	l.RegisterExporter("genesyscloud_auth_division", AuthDivisionExporter())
 	l.RegisterExporter("genesyscloud_group", GroupExporter())
 	l.RegisterExporter("genesyscloud_idp_adfs", IdpAdfsExporter())
