+++ genesyscloud/resource_genesyscloud_init_test.go
@@ -33,7 +33,6 @@ func (r *registerTestInstance) registerTestResources() {
 	providerResources["genesyscloud_routing_queue"] = ResourceRoutingQueue()
 	providerResources["genesyscloud_location"] = ResourceLocation()
 	providerResources["genesyscloud_architect_schedules"] = ResourceArchitectSchedules()
-	providerResources["genesyscloud_architect_user_prompt"] = ResourceArchitectUserPrompt()
 	providerResources["genesyscloud_auth_division"] = ResourceAuthDivision()
 	providerResources["genesyscloud_group"] = ResourceGroup()
 	providerResources["genesyscloud_idp_adfs"] = ResourceIdpAdfs()
@@ -81,7 +80,6 @@ func (r *registerTestInstance) registerTestDataSources() {
 	providerDataSources["genesyscloud_location"] = DataSourceLocation()
 	providerDataSources["genesyscloud_auth_division_home"] = DataSourceAuthDivisionHome()
 	providerDataSources["genesyscloud_architect_schedules"] = DataSourceSchedule()
-	providerDataSources["genesyscloud_architect_user_prompt"] = dataSourceUserPrompt()
 	providerDataSources["genesyscloud_auth_division"] = dataSourceAuthDivision()
 	providerDataSources["genesyscloud_auth_division_home"] = DataSourceAuthDivisionHome()
 	providerDataSources["genesyscloud_group"] = DataSourceGroup()
