+++ genesyscloud/tfexporter/resource_genesyscloud_tf_export_test.go
@@ -13,6 +13,7 @@ import (
 	"strings"
 	gcloud "terraform-provider-genesyscloud/genesyscloud"
 	"terraform-provider-genesyscloud/genesyscloud/architect_flow"
+	userPrompt "terraform-provider-genesyscloud/genesyscloud/architect_user_prompt"
 	obContactList "terraform-provider-genesyscloud/genesyscloud/outbound_contact_list"
 	"terraform-provider-genesyscloud/genesyscloud/provider"
 	resourceExporter "terraform-provider-genesyscloud/genesyscloud/resource_exporter"
@@ -989,7 +990,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		exportTestDir    = "../.terraform" + uuid.NewString()
 	)
 
-	userPromptAsset := gcloud.UserPromptResourceStruct{
+	userPromptAsset := userPrompt.UserPromptResourceStruct{
 		Language:        userPromptResourceLanguage,
 		Tts_string:      util.NullValue,
 		Text:            strconv.Quote(userPromptResourceText),
@@ -997,7 +998,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		FileContentHash: userResourcePromptFilename1,
 	}
 
-	userPromptAsset2 := gcloud.UserPromptResourceStruct{
+	userPromptAsset2 := userPrompt.UserPromptResourceStruct{
 		Language:        userPromptResourceLanguage2,
 		Tts_string:      util.NullValue,
 		Text:            strconv.Quote(userPromptResourceText2),
@@ -1005,8 +1006,8 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		FileContentHash: userResourcePromptFilename2,
 	}
 
-	userPromptResources := []*gcloud.UserPromptResourceStruct{&userPromptAsset}
-	userPromptResources2 := []*gcloud.UserPromptResourceStruct{&userPromptAsset, &userPromptAsset2}
+	userPromptResources := []*userPrompt.UserPromptResourceStruct{&userPromptAsset}
+	userPromptResources2 := []*userPrompt.UserPromptResourceStruct{&userPromptAsset, &userPromptAsset2}
 
 	defer os.RemoveAll(exportTestDir)
 
@@ -1015,7 +1016,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 		ProviderFactories: provider.GetProviderFactories(providerResources, providerDataSources),
 		Steps: []resource.TestStep{
 			{
-				Config: gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
+				Config: userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
@@ -1034,7 +1035,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 					[]string{
 						strconv.Quote("genesyscloud_architect_user_prompt." + userPromptResourceId),
 					},
-				) + gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
+				) + userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
@@ -1051,7 +1052,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 			},
 			// Update to two resources with separate audio files
 			{
-				Config: gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
+				Config: userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
@@ -1070,7 +1071,7 @@ func TestAccResourceTfExportUserPromptExportAudioFile(t *testing.T) {
 					[]string{
 						strconv.Quote("genesyscloud_architect_user_prompt." + userPromptResourceId),
 					},
-				) + gcloud.GenerateUserPromptResource(&gcloud.UserPromptStruct{
+				) + userPrompt.GenerateUserPromptResource(&userPrompt.UserPromptStruct{
 					ResourceID:  userPromptResourceId,
 					Name:        userPromptName,
 					Description: strconv.Quote(userPromptDescription),
